requirements.txt
accelerate==0.25.0
bark_ssg==1.3.4
bert_embedding==1.0.1
bitsandbytes==0.42.0
datasets==2.12.0
diffusers==0.25.0
gensim==4.3.2
gpustat==1.1.1
h5py==3.10.0
joblib==1.3.2
librosa==0.10.1
networkx==3.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
peft==0.7.1
Pillow==10.0.1
Pillow==10.2.0
preprocessor==1.1.3
psutil==5.9.7
pytorch_pretrained_bert==0.6.2
scikit_learn==1.3.2
scipy==1.11.4
torch==1.12.1
torchtext==0.6.0
torchvggish==0.2
torchvision==0.13.1
tqdm==4.65.0
transformers==4.36.2
wandb==0.16.2
